\name{mMCPcounter_signatures}
\alias{mMCPcounter_signatures}
\docType{data}
\title{
Signatures for mMCP-counter
}
\description{
Set of gene signatures for the various populations measured by mMCP-counter
}
\usage{data("mMCPcounter_signatures")}
\examples{
data("mMCPcounter_signatures")
}
\keyword{datasets}
