\name{mMCPcounter_signatures_GCRm38}
\alias{mMCPcounter_signatures_GCRm38}
\docType{data}
\title{
Signatures for mMCP-counter with version GCRm38 of the mouse genome.
}
\description{
Set of gene signatures for the various populations measured by mMCP-counter
}
\usage{data("mMCPcounter_signatures_GCRm38")}
\examples{
data("mMCPcounter_signatures_GCRm38")
}
\keyword{datasets}
