\name{mMCPcounter_signatures_GCRm39}
\alias{mMCPcounter_signatures_GCRm39}
\docType{data}
\title{
Signatures for mMCP-counter with version GCRm39 of the mouse genome.
}
\description{
Set of gene signatures for the various populations measured by mMCP-counter
}
\usage{data("mMCPcounter_signatures_GCRm39")}
\examples{
data("mMCPcounter_signatures_GCRm39")
}
\keyword{datasets}
